requirements.txt
eth-brownie>=1.12.0,<2.0.0